Shifted to two stage bootloader
The current bootloader didn't allow any more code to be added since it was almost 500+ bytes.
The memory detection module has to be added to bootloader so we needed the two stage bootloader. The other option was to highly compress the boot sector code, which didn't provide much of benefits and the two stage bootloader was not that difficult.